refactor!: move announce popover
